Fixes GDT Bug

+++ protected-mode/src/gdt/gdt.h
@@ -4,7 +4,7 @@
 struct gdt {
     uint16_t segment;
     uint16_t base_first;
-    uint16_t base;
+    uint8_t base;
     uint8_t access;
     uint8_t high_flags;
     uint8_t base_24_31_bits;
